+++ changelog.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * `convex-base`
-  - `Convex.PlutusLedger.transAddressInEra`
+  - `Convex.PlutusLedger`: `transAddressInEra`, prisms and isos for the `Interval` type
   - `Convex.Utxos.singleton`
   - `Convex.Utxos`: Added redeemer and ex units to `RemoveUtxoEvent`
   - `Convex.Lenses`: Added `_TxOutDatumInline`, `_TxOutDatumInTx`, `_ScriptData`, `_PlutusScriptWitness`, `_TxExtraKeyWitnesses`, prisms for validity intervals